requirements.txt
PyTorch==1.5.0
torchvision==0.6.0
BoTorch==0.2.5
timm==0.1.26